outputs/generated_repo/car-website-showing-all-honda-cars-sold-in-pakistan/main.py
```python
# Import required libraries
import os
import sys
from datetime import datetime
from github import Github
from github import InputFileContent

# Define constants
GITHUB_TOKEN = "your_github_token"
GITHUB_USERNAME = "your_github_username"
REPO_NAME = "car-website-honda-pakistan"

# Define a function to generate project specification
def generate_project_specification():
    """
    Generates a project specification for the car website.
    """
    specification = f"""
# Car Website Project Specification

## Overview
The car website is an open-source project that aims to display all Honda cars sold in Pakistan.

## Features
- Display a list of all Honda cars sold in Pakistan
- Provide details of each car, including model, year, price, and features
- Allow users to filter cars by model, year, and price

## Requirements
- Python 3.8+
- Flask 2.0+
- GitHub repository for version control

## Timeline
- Week 1: Set up GitHub repository and project scaffolding
- Week 2-3: Develop the car website using Flask
- Week 4: Test and deploy the website
"""
    return specification

# Define a function to scaffold the repository
def scaffold_repository():
    """
    Scaffolds the GitHub repository for the car website project.
    """
    # Create a new GitHub repository
    g = Github(GITHUB_TOKEN)
    user = g.get_user(GITHUB_USERNAME)
    repo = user.create_repo(REPO_NAME, description="Car website showing all Honda cars sold in Pakistan")

    # Create a new branch for the project
    repo.create_git_ref(ref="refs/heads/development", sha=repo.get_branch("main").commit.sha)

    # Create a new file for the project specification
    spec_file = InputFileContent(content=generate_project_specification())
    repo.create_file(path="PROJECT_SPECIFICATION.md", message="Initial project specification", content=spec_file, branch="development")

# Define a function to draft documentation
def draft_documentation():
    """
    Drafts documentation for the car website project.
    """
    documentation = f"""
# Car Website Documentation

## Introduction
The car website is an open-source project that aims to display all Honda cars sold in Pakistan.

## Getting Started
1. Clone the repository: `git clone https://github.com/{GITHUB_USERNAME}/{REPO_NAME}.git`
2. Install dependencies: `pip install -r requirements.txt`
3. Run the application: `python main.py`

## API Endpoints
- `/cars`: Returns a list of all Honda cars sold in Pakistan
- `/cars/{model}`: Returns details of a specific car model
"""
    return documentation

# Define a function to set up GitHub workflow
def setup_github_workflow():
    """
    Sets up a GitHub workflow for the car website project.
    """
    # Create a new file for the GitHub workflow
    workflow_file = InputFileContent(content="""
name: Car Website Workflow

on:
  push:
    branches:
      - main

jobs:
  build:
    runs-on: ubuntu-latest
    steps:
      - name: Checkout code
        uses: actions/checkout@v2
      - name: Install dependencies
        run: pip install -r requirements.txt
      - name: Run application
        run: python main.py
""")
    repo = Github(GITHUB_TOKEN).get_repo(f"{GITHUB_USERNAME}/{REPO_NAME}")
    repo.create_file(path=".github/workflows/main.yml", message="Initial GitHub workflow", content=workflow_file, branch="development")

# Define the main application logic
def main():
    print("Car Website Project")
    print("--------------------")
    print("1. Generate project specification")
    print("2. Scaffold repository")
    print("3. Draft documentation")
    print("4. Set up GitHub workflow")
    print("5. Exit")
    
    while True:
        choice = input("Enter your choice: ")
        
        if choice == "1":
            print(generate_project_specification())
        elif choice == "2":
            scaffold_repository()
            print("Repository scaffolded successfully!")
        elif choice == "3":
            print(draft_documentation())
        elif choice == "4":
            setup_github_workflow()
            print("GitHub workflow set up successfully!")
        elif choice == "5":
            break
        else:
            print("Invalid choice. Please try again.")

if __name__ == "__main__":
    main()
```

**Note:** You need to replace `"your_github_token"` and `"your_github_username"` with your actual GitHub token and username.

**Example Use Case:**

1. Run the application: `python main.py`
2. Choose an option from the menu:
	* Generate project specification: `1`
	* Scaffold repository: `2`
	* Draft documentation: `3`
	* Set up GitHub workflow: `4`
	* Exit: `